requirements.txt
torch
torchvision
timm
mediapipe
opencv-python
grad-cam
scikit-learn
matplotlib
seaborn
flask
gradio
kagglehub
numpy
Pillow
